+++ cmd/hidden_lake/applications/messenger/internal/config/config.go
@@ -219,14 +219,3 @@ func (p *sLogging) HasWarn() bool {
 func (p *sLogging) HasErro() bool {
 	return (*p)[2]
 }
-
-func (p *SConfig) GetSecretKeys() map[string]string {
-	p.fMutex.Lock()
-	defer p.fMutex.Unlock()
-
-	result := make(map[string]string)
-	for k, v := range p.FSecretKeys {
-		result[k] = v
-	}
-	return result
-}